Move processRoot function into the utils.js file
