+++ lib/render.js
@@ -2,7 +2,7 @@ var extend = require('deepmerge');
 var fm = require('front-matter');
 var path = require('path');
 var through = require('through2');
-var processRoot = require('./processRoot');
+var processRoot = require('./utils').processRoot;
 
 module.exports = function() {
   return through.obj(render.bind(this));
